requirements.txt
alembic==1.6.5
appdirs==1.4.4
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
beautifulsoup4==4.9.3
bidict==0.21.2
black==20.8b1
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.3.0
Deprecated==1.2.12
distlib==0.3.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
emails==0.6
filelock==3.0.12
Flask==2.0.1
Flask-Admin==1.5.8
flask-admin.py==0.1.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-BS4==5.0.0.1
flask-emails==0.4.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
flask-msearch==0.2.9.2
flask-peewee==3.0.4
Flask-Script==2.0.6
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.15.1
fpdf==1.7.2
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.1
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Js2Py==0.71
login==0.0.6
lxml==4.6.3
mailclient==0.1.1
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
pathspec==0.9.0
peewee==3.14.4
pep8==1.7.1
Pillow==8.3.1
pipupgrade==1.9.0
pkg_resources==0.0.0
platformdirs==2.2.0
premailer==3.10.0
pycodestyle==2.7.0
pycparser==2.20
pyjsparser==2.7.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
regex==2021.8.3
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
textmagic==2.0.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==3.0
uni-slugify==0.1.4
Unidecode==1.2.0
urllib3==1.26.6
vext==0.7.6
vext.gi==0.7.4
virtualenv==20.7.2
visitor==0.1.3
vonage==2.5.5
Werkzeug==2.0.1
Whoosh==2.7.4
wrapt==1.12.1
wtf-peewee==3.0.2
WTForms==2.3.3
